Moved the validate trait into the command namespace
Validate is only used from the commands.
Tidy the root namespace.
