+++ src/Command/Ignore.php
@@ -6,7 +6,6 @@ namespace Phlib\Beanstalk\Command;
 
 use Phlib\Beanstalk\Connection\Socket;
 use Phlib\Beanstalk\Exception\CommandException;
-use Phlib\Beanstalk\ValidateTrait;
 
 /**
  * @package Phlib\Beanstalk
