+++ src/Command/Put.php
@@ -8,7 +8,6 @@ use Phlib\Beanstalk\Connection\Socket;
 use Phlib\Beanstalk\Exception\BuriedException;
 use Phlib\Beanstalk\Exception\CommandException;
 use Phlib\Beanstalk\Exception\DrainingException;
-use Phlib\Beanstalk\ValidateTrait;
 
 /**
  * @package Phlib\Beanstalk
